# openapi-spec/openapi-spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Vehicle Locator API",
        "description": "API to track and manage vehicle locations",
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "https://vehicle-api-secured.42crunch.test"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/vehicles": {
            "get": {
                "summary": "Get all vehicles",
                "operationId": "getVehicles",
                "responses": {
                    "200": {
                        "description": "A list of all vehicles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/VehicleList"
                                    },
                                    "maxItems": 10
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Default",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorMessage"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add a new vehicle",
                "operationId": "postVehicle",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehicleInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehicle created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehicle"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorMessage"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Vehicle ID already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorMessage"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Default",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorMessage"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorMessage": {
                "required": [
                    "message"
                ],
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "message": {
                        "maxLength": 255,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.\\-@:,;]{1,255}$",
                        "type": "string"
                    }
                }
            },
...
            "Vehicle": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "MongoDB ID"
                    },
                    "id": {
                        "type": "string",
                        "description": "Vehicle ID",
                        "format": "uuid",
                        "example": "a83a29f5-0d63-46f2-8f2e-44c2f1d2e07e",
                        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
                        "minLength": 36,
                        "maxLength": 36,
                        "x-42c-format": "o:uuid"
                    },
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "description": "Latitude",
                        "minimum": -90,
                        "maximum": 90
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "description": "Longitude",
                        "minimum": -180,
                        "maximum": 180
                    },
                    "vin": {
                        "type": "string",
                        "description": "Vehicle Identification Number",
                        "format": "o:Vehicle-Locator-API:VIN",
                        "example": "SALVA2AE4EH877322",
                        "pattern": "^[A-Z0-9]{17}$",
                        "minLength": 17,
                        "maxLength": 17,
                        "x-42c-format": "o:Vehicle-Locator-API:VIN"
                    },
                    "make": {
                        "type": "string",
                        "description": "Vehicle make",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$"
                    },
                    "model": {
                        "type": "string",
                        "description": "Vehicle model",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$"
                    },
                    "year": {
                        "type": "integer",
                        "description": "Vehicle year",
                        "minimum": 1000,
                        "maximum": 3000
                    },
                    "color": {
                        "type": "string",
                        "description": "Vehicle color",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "description": "Vehicle status"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Last update date",
                        "example": "2021-08-30T23:01:14.274085491+15:55",
                        "pattern": "^[0-9]{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12][0-9]|3[01])T(?:[01][0-9]|2[0-3]):[0-5][0-9]:(?:[0-5][0-9]|60)(?:\\.[0-9]{1,15})?(?:Z|[+\\-](?:[01][0-9]|2[0-3]):[0-5][0-9])$",
                        "minLength": 20,
                        "maxLength": 38,
                        "x-42c-format": "o:date-time"
                    }
                },
                "required": [
                    "id",
                    "lat",
                    "lng"
                ]
            },
            "VehicleList": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Vehicle ID",
                        "format": "uuid",
                        "example": "a83a29f5-0d63-46f2-8f2e-44c2f1d2e07e",
                        "pattern": "^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$",
                        "minLength": 36,
                        "maxLength": 36,
                        "x-42c-format": "o:uuid"
                    },
                    "vin": {
                        "type": "string",
                        "description": "Vehicle Identification Number",
                        "format": "o:Vehicle-Locator-API:VIN",
                        "example": "SALVA2AE4EH877322",
                        "pattern": "^[A-Z0-9]{17}$",
                        "minLength": 17,
                        "maxLength": 17,
                        "x-42c-format": "o:Vehicle-Locator-API:VIN"
                    },
                    "make": {
                        "type": "string",
                        "description": "Vehicle make",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$"
                    },
                    "model": {
                        "type": "string",
                        "description": "Vehicle model",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$"
                    },
                    "year": {
                        "type": "integer",
                        "description": "Vehicle year",
                        "minimum": 1000,
                        "maximum": 3000
                    },
                    "color": {
                        "type": "string",
                        "description": "Vehicle color",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive"
                        ],
                        "description": "Vehicle status"
                    },
                    "lastUpdated": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Last update date",
                        "example": "2021-08-30T23:01:14.274085491+15:55",
                        "pattern": "^[0-9]{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12][0-9]|3[01])T(?:[01][0-9]|2[0-3]):[0-5][0-9]:(?:[0-5][0-9]|60)(?:\\.[0-9]{1,15})?(?:Z|[+\\-](?:[01][0-9]|2[0-3]):[0-5][0-9])$",
                        "minLength": 20,
                        "maxLength": 38,
                        "x-42c-format": "o:date-time"
                    }
                },
                "required": [
                    "id",
                    "vin"
                ]
            },
            "VehicleInput": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                    "vin": {
                        "type": "string",
                        "format": "o:Vehicle-Locator-API:VIN",
                        "example": "SALVA2AE4EH877322",
                        "pattern": "^[A-Z0-9]{17}$",
                        "minLength": 17,
                        "maxLength": 17,
                        "x-42c-format": "o:Vehicle-Locator-API:VIN"
                    },
                    "lat": {
                        "type": "number",
                        "format": "float",
                        "minimum": -90,
                        "maximum": 90
                    },
                    "lng": {
                        "type": "number",
                        "format": "float",
                        "minimum": -180,
                        "maximum": 180
                    },
                    "make": {
                        "type": "string",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$",
                        "description": "Vehicle make"
                    },
                    "model": {
                        "type": "string",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$",
                        "description": "Vehicle model"
                    },
                    "year": {
                        "type": "integer",
                        "description": "Vehicle year",
                        "minimum": 1000,
                        "maximum": 3000,
                        "format": "int32"
                    },
                    "color": {
                        "type": "string",
                        "maxLength": 30,
                        "minLength": 1,
                        "pattern": "^[\\w\\s\\.]{1,30}$",
                        "description": "Vehicle color"
                    }
                },
                "required": [
                    "vin",
                    "lat",
                    "lng",
                    "make",
                    "model",
                    "year",
                    "color"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}